Commit 027 - Revisando o projeto

+++ backend-springboot/src/main/java/br/com/compasso/backend/controller/EstadoController.java
@@ -1,13 +1,21 @@
 package br.com.compasso.backend.controller;
 
+import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
+
+import javax.swing.text.html.Option;
+
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
+import br.com.compasso.backend.model.CidadeModel;
 import br.com.compasso.backend.model.EstadoModel;
 import br.com.compasso.backend.repository.EstadoRepository;
 
@@ -25,39 +33,67 @@ public class EstadoController {
 	
 	@CrossOrigin
 	@RequestMapping(value = "/estadoNome/{nome}", method = RequestMethod.GET)
-	public EstadoModel GetByEstado(@PathVariable(value = "nome") String nome) {
-		return estadoRepository.findByNome(nome);
+	public ResponseEntity<EstadoModel> GetByEstado(@PathVariable(value = "nome") String nome) {
+		EstadoModel estado = estadoRepository.findByNome(nome);
+		if (estado != null) {
+			return ResponseEntity.ok(estado);
+		}
+		
+		return ResponseEntity.notFound().build();
 	}
 	
 	@CrossOrigin
 	@RequestMapping(value = "/estadoUf/{uf}", method = RequestMethod.GET)
-	public EstadoModel GetByUf(@PathVariable(value = "uf") String uf) {
-		return estadoRepository.findByUf(uf);
+	public ResponseEntity<EstadoModel> GetByUf(@PathVariable(value = "uf") String uf) {
+		EstadoModel estado = estadoRepository.findByUf(uf);
+		if (estado != null) {
+			return ResponseEntity.ok(estado);
+		}
+		
+		return ResponseEntity.notFound().build();
 	}
 	
 	@CrossOrigin
 	@RequestMapping(value = "/estado", method = RequestMethod.POST)
-	public EstadoModel estadoCreate(@RequestBody EstadoModel estado) {
-		EstadoModel estadoModel = new EstadoModel();
-		estadoModel.setNome(estado.getNome());
-		estadoModel.setUf(estado.getUf());
-		
-		return estadoRepository.save(estadoModel);
+	public ResponseEntity<EstadoModel> estadoCreate(@RequestBody EstadoModel estado) {
+		EstadoModel buscaEstado = estadoRepository.findByNome(estado.getNome());
+		if (buscaEstado != null) {
+			return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+		} else {
+			return ResponseEntity.ok(estadoRepository.save(estado));
+		}
 	}
 	
 	@CrossOrigin
 	@RequestMapping(value = "/estado/{id_estado}", method = RequestMethod.PUT)
-	public EstadoModel estadoUpdate(@PathVariable(value = "id_estado") long estadoId, @RequestBody EstadoModel estado) {
-		EstadoModel estadoModel = estadoRepository.findById(estadoId);
-		estadoModel.setNome(estado.getNome());
-		estadoModel.setUf(estado.getUf());
+	public ResponseEntity<EstadoModel> estadoUpdate(@PathVariable(value = "id_estado") Long estadoId, @RequestBody EstadoModel estado) {
+		/*Optional<EstadoModel> buscaEstado = estadoRepository.findById(estadoId);
+		buscaEstado.se
+		if (buscaEstado.isPresent()) {	
+			return ResponseEntity.ok(estadoRepository.save(estado));
+		} else {
+			return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+		}*/
 		
-		return estadoRepository.save(estadoModel);
+	//	estado.setCidades(new Optional<CidadeModel>());
+		
+		EstadoModel buscaEstado = estadoRepository.findById(estadoId).get();
+
+		if (buscaEstado != null) {	
+			return ResponseEntity.ok(estadoRepository.save(estado));
+		} else {
+			return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+		}
 	}
 	
 	@CrossOrigin
 	@RequestMapping(value = "/estado/{id_estado}", method = RequestMethod.DELETE)
-	public EstadoModel estadoDelete(@PathVariable(value = "id_estado") long estadoId) {
-		return estadoRepository.deleteById(estadoId);
+	public ResponseEntity<CidadeModel> cidadeDelete(@PathVariable(value = "id_estado") Long estadoId) {
+		Optional<EstadoModel> cidade = estadoRepository.findById(estadoId);
+		if (cidade.isPresent()) {
+			estadoRepository.deleteById(estadoId);
+			return ResponseEntity.status(HttpStatus.OK).build();
+		} 
+		return ResponseEntity.notFound().build();
 	}
 }
